Completely removed the EqC and LGEC data types
Providing different data types for case-insensitive (really,
X-insensitive for a caller-provided X) strings implies (1) breaking
the expectation of what `==`, `<`, `>`, etc. do for strings, (2)
additional API complexity, and (3) additional code maintenance
(testing and documentation).  In practice, it should suffice for a
caller to canonicalize a string before invoking NewEq or NewLGE.

If someone suggests a practical use case for EqC and LGEC, they can
always be restored.